Move deploy-worker command from server CLI to client CLI
The deploy-worker command is pure subprocess orchestration with no
database or API interaction, so it belongs in the client CLI (docverse)
rather than the server CLI (docverse-admin) which eagerly loads server
configuration at import time.
