+++ client/src/docverse/client/_cli.py
@@ -3,6 +3,9 @@
 from __future__ import annotations
 
 import asyncio
+import json
+import re
+import shutil
 import subprocess
 import sys
 from pathlib import Path
@@ -146,6 +149,117 @@ def upload(  # noqa: PLR0913
             verbose=verbose,
         )
     )
+
+
+@main.command()
+@click.option(
+    "--docverse-repo",
+    required=True,
+    type=click.Path(exists=True, file_okay=False, path_type=Path),
+    help="Path to the docverse monorepo root.",
+)
+@click.option(
+    "--deployments-repo",
+    required=True,
+    type=click.Path(exists=True, file_okay=False, path_type=Path),
+    help="Path to local docverse-cloudflare-deployments checkout.",
+)
+@click.option(
+    "--env",
+    "wrangler_env",
+    required=True,
+    help="Wrangler environment name (e.g., dev, production).",
+)
+@click.option(
+    "--dry-run",
+    is_flag=True,
+    default=False,
+    help="Build the worker bundle without deploying.",
+)
+def deploy_worker(
+    *,
+    docverse_repo: Path,
+    deployments_repo: Path,
+    wrangler_env: str,
+    dry_run: bool,
+) -> None:
+    """Pack and deploy the Cloudflare Worker."""
+    if not re.match(r"^[a-zA-Z0-9_-]+$", wrangler_env):
+        msg = (
+            f"Invalid environment name {wrangler_env!r}: "
+            "must contain only alphanumeric characters, hyphens, "
+            "and underscores"
+        )
+        raise click.ClickException(msg)
+
+    worker_dir = docverse_repo.resolve() / "cloudflare-worker"
+    if not worker_dir.is_dir():
+        msg = f"cloudflare-worker/ not found in {docverse_repo.resolve()}"
+        raise click.ClickException(msg)
+
+    deployments_repo = deployments_repo.resolve()
+    dest_dir = deployments_repo / "worker"
+    dest_dir.mkdir(exist_ok=True)
+
+    # Phase 1: npm pack
+    click.echo(f"Packing cloudflare worker from {worker_dir}")
+    try:
+        pack_result = subprocess.run(
+            ["npm", "pack", "--json"],  # noqa: S607
+            check=True,
+            capture_output=True,
+            text=True,
+            cwd=str(worker_dir),
+        )
+    except subprocess.CalledProcessError as exc:
+        msg = f"npm pack failed: {exc.stderr}"
+        raise click.ClickException(msg) from exc
+    try:
+        tarball_name = json.loads(pack_result.stdout)[0]["filename"]
+    except (json.JSONDecodeError, KeyError, IndexError) as exc:
+        msg = f"Failed to parse npm pack output: {pack_result.stdout!r}"
+        raise click.ClickException(msg) from exc
+
+    # Phase 2: copy and unpack into deployments repo
+    click.echo(f"Unpacking worker into {dest_dir}")
+    shutil.copy2(worker_dir / tarball_name, dest_dir / tarball_name)
+    try:
+        subprocess.run(  # noqa: S603
+            ["tar", "xzf", tarball_name, "--strip-components=1"],  # noqa: S607
+            check=True,
+            cwd=str(dest_dir),
+        )
+    except subprocess.CalledProcessError as exc:
+        msg = f"Failed to unpack worker tarball: {exc}"
+        raise click.ClickException(msg) from exc
+    finally:
+        for tgz in dest_dir.glob("*.tgz"):
+            tgz.unlink()
+    (worker_dir / tarball_name).unlink(missing_ok=True)
+
+    # Phase 3: wrangler deploy
+    wrangler_cmd = [
+        "npx",
+        "wrangler",
+        "deploy",
+        "--env",
+        wrangler_env,
+    ]
+    if dry_run:
+        outdir = deployments_repo / "dist"
+        wrangler_cmd.extend(["--dry-run", f"--outdir={outdir}"])
+    click.echo(f"Deploying worker (env={wrangler_env}, dry_run={dry_run})")
+    try:
+        subprocess.run(  # noqa: S603
+            wrangler_cmd,
+            check=True,
+            cwd=str(deployments_repo),
+        )
+    except subprocess.CalledProcessError as exc:
+        msg = f"wrangler deploy failed: {exc}"
+        raise click.ClickException(msg) from exc
+
+    click.echo(f"Worker deployed successfully (env={wrangler_env})")
 
 
 def _detect_git_ref() -> str:
